+++ src/main/java/bus/ThongTinSDBUS.java
@@ -247,7 +247,7 @@ public class ThongTinSDBUS {
         return response;
     }
     
-    public Response<Boolean> delete(int id) {
+    public Response<Boolean> delete(long id) {
         Response<Boolean> response = new Response<>();
         try {
