+++ src/main/java/view/page/ThongTinSDPage.java
@@ -69,7 +69,7 @@ public class ThongTinSDPage extends javax.swing.JPanel {
         event = new TableActionEvent() {
             @Override
             public void onEdit(int row) {
-                int id =Integer.parseInt(table.getModel().getValueAt(row, 0).toString());
+                long id =Integer.parseInt(table.getModel().getValueAt(row, 0).toString());
                 isUpdating=true;
                 oldId = id;
                 renderDetail(id);
@@ -80,7 +80,7 @@ public class ThongTinSDPage extends javax.swing.JPanel {
                 if(Ok!=JOptionPane.YES_OPTION)
                     return;
                 DefaultTableModel model = (DefaultTableModel) table.getModel();
-                int id = Integer.parseInt(model.getValueAt(row, 0).toString());
+                long id = Integer.parseInt(model.getValueAt(row, 0).toString());
                 Response response = thongTinSDBUS.delete(id);
                 if(response.getStatus() == ResponseStatus.FAILURE){
                     JOptionPane.showMessageDialog(null, response.getMessage(), "Warning", JOptionPane.WARNING_MESSAGE);
@@ -138,7 +138,7 @@ public class ThongTinSDPage extends javax.swing.JPanel {
             model.addRow(row);
         }
     }
-    public void renderDetail(int id){
+    public void renderDetail(long id){
 //        Response<ThongTin> response = thongTinSDBUS.getOne("MaTV",id);
 //        if(response.getStatus() == ResponseStatus.FAILURE){
 //            JOptionPane.showMessageDialog(null, response.getMessage(), "Warning", JOptionPane.WARNING_MESSAGE);
@@ -984,7 +984,7 @@ public class ThongTinSDPage extends javax.swing.JPanel {
             return null;
         }
     }
-    private int oldId;
+    private long oldId;
     private boolean isUpdating;
     private customDialog checkinForm,issueForm,returnForm;
     // Variables declaration - do not modify//GEN-BEGIN:variables
